fix(engine): widen solution counters to size_t; add frontier-DP prototype
The counters overflowed at 2^31. runningCount was a 32-bit int, and nine of
the class trees have budgets whose counts exceed INT_MAX -- shaman_class_
elemental reaches 37,296,642,700 at its best budget.

The failure was not a wrong number, it was a silent truncation. runningCount
wrapped to INT_MIN, and because the guard compares it against a size_t the
negative value promoted to a huge unsigned, tripping the guard and stopping
enumeration early. Observed on shaman_class_elemental at 24 points:

    Tree 0: 18446744071562067968 combinations (INCOMPLETE: safety guard
    triggered)          <- 2^64 - 2^31, i.e. INT_MIN reinterpreted

This was reachable only after the previous commit started incrementing
runningCount on the filtered path; before that the counter was never
incremented there, so the overflow could not occur. Adding the guard was
correct, but it collided with the 32-bit counter.

runningCount and the allCombinations accumulators are now size_t, which also
removes the signed/unsigned comparison in the guard check. TreeDAGInfo fields
were already size_t. Verified no regression: 223 / 17,541 / 71,030 on
druid_restoration at 10/15/20 points, on both MSVC v143 and gcc 12. The native
GUI only touches TreeDAGInfo::allCombinations, whose type is unchanged.

Also adds tools/frontier-dp/: a verified prototype that counts valid builds
with dynamic programming instead of enumerating them. Cost is polynomial in
tree size rather than proportional to the number of solutions.

    druid_restoration, 30 points        engine 42.0 s   DP  --
    druid_restoration, all budgets      30 runs         DP  0.10 s
    all 78 presets, all budgets         --              DP  3.7 s

Counts are identical to the engine on every tree tested: 11 trees across 6
classes, class and spec trees, plus 223 / 17,541 / 71,030 / 305,286,987 on
druid_restoration. Peak DP state count across every tree: 8,859.

It does not yet handle hero talents, tiered level-gated ranks, or include/
one-of filters, and it counts rather than lists -- enumeration remains the
right tool for listing builds that match filters.

+++ Engine/src/TreeSolver.cpp
@@ -75,7 +75,7 @@ namespace Engine {
         }
         auto t1 = std::chrono::high_resolution_clock::now();
         //this is used for safeguarding solving process for trees that are too big
-        int runningCount = 0;
+        size_t runningCount = 0;
         for (int i = 0; i < possibleTalents.size(); i++) {
             //only start with root nodes that have points required == 0, prevents from starting at root nodes that might come later in the tree (e.g. druid wild charge)
             if (sortedTreeDAG.sortedTalents[possibleTalents[i].first]->pointsRequired == 0)
@@ -128,7 +128,7 @@ namespace Engine {
         std::vector<std::pair<int, int>> possibleTalents,
         const TreeDAGInfo& sortedTreeDAG,
         std::vector<SIND>& combinations,
-        int& runningCount,
+        size_t& runningCount,
         bool& safetyGuardTriggered
     ) {
         /*
@@ -307,7 +307,7 @@ namespace Engine {
         }
         auto t1 = std::chrono::high_resolution_clock::now();
         //this is used for safeguarding solving process for trees that are too big
-        int runningCount = 0;
+        size_t runningCount = 0;
         for (int i = 0; i < possibleTalents.size(); i++) {
             //only start with root nodes that have points required == 0, prevents from starting at root nodes that might come later in the tree (e.g. druid wild charge)
             if (sortedTreeDAG.sortedTalents[possibleTalents[i].first]->pointsRequired == 0)
@@ -337,7 +337,7 @@ namespace Engine {
         vec2d<SIND> allCombinationsVector;
         allCombinationsVector.push_back(combinations);
         sortedTreeDAG.allCombinations = allCombinationsVector;
-        sortedTreeDAG.resultCount = static_cast<size_t>(runningCount);
+        sortedTreeDAG.resultCount = runningCount;
         sortedTreeDAG.elapsedTime = ms_double.count() / 1000.0;
         inProgress = false;
 
@@ -366,7 +366,7 @@ namespace Engine {
         std::vector<std::pair<int, int>> possibleTalents,
         const TreeDAGInfo& sortedTreeDAG,
         std::vector<SIND>& combinations,
-        int& runningCount,
+        size_t& runningCount,
         bool& safetyGuardTriggered,
         SIND& includeFilter,
         SIND& excludeFilter,
@@ -468,7 +468,7 @@ namespace Engine {
             throw std::logic_error("Number of talents exceeds 64, need different indexing type instead of uint64");
         vec2d<SIND> combinations;
         combinations.resize(talentPoints);
-        std::vector<int> allCombinations;
+        std::vector<size_t> allCombinations;
         allCombinations.resize(talentPoints, 0);
 
         //iterate through all possible combinations in order:
@@ -483,7 +483,7 @@ namespace Engine {
         }
         auto t1 = std::chrono::high_resolution_clock::now();
         //this is used for safeguarding solving process for trees that are too big
-        int runningCount = 0;
+        size_t runningCount = 0;
         for (int i = 0; i < possibleTalents.size(); i++) {
             //only start with root nodes that have points required == 0, prevents from starting at root nodes that might come later in the tree (e.g. druid wild charge)
             //if (possibleTalents[i].second == 0)
@@ -522,8 +522,8 @@ namespace Engine {
         std::vector<std::pair<int, int>> possibleTalents,
         const TreeDAGInfo& sortedTreeDAG,
         vec2d<SIND>& combinations,
-        std::vector<int>& allCombinations,
-        int& runningCount,
+        std::vector<size_t>& allCombinations,
+        size_t& runningCount,
         bool& safetyGuardTriggered
     ) {
         if (runningCount >= sortedTreeDAG.safetyGuard || safetyGuardTriggered) {
@@ -585,7 +585,7 @@ namespace Engine {
             throw std::logic_error("Number of talents exceeds 64, need different indexing type instead of uint64");
         vec2d<std::pair<SIND, int>> combinations;
         combinations.resize(talentPoints);
-        std::vector<int> allCombinations;
+        std::vector<size_t> allCombinations;
         allCombinations.resize(talentPoints, 0);
 
         //iterate through all possible combinations in order:
@@ -601,7 +601,7 @@ namespace Engine {
         }
         auto t1 = std::chrono::high_resolution_clock::now();
         //this is used for safeguarding solving process for trees that are too big
-        int runningCount = 0;
+        size_t runningCount = 0;
         bool safetyGuardTriggered = false;
         for (int i = 0; i < possibleTalents.size(); i++) {
             //only start with root nodes that have points required == 0, prevents from starting at root nodes that might come later in the tree (e.g. druid wild charge)
@@ -632,8 +632,8 @@ namespace Engine {
         std::vector<std::pair<int, int>> possibleTalents,
         const TreeDAGInfoLegacy& sortedTreeDAG,
         vec2d<std::pair< SIND, int>>& combinations,
-        std::vector<int>& allCombinations,
-        int& runningCount,
+        std::vector<size_t>& allCombinations,
+        size_t& runningCount,
         bool& safetyGuardTriggered
     ) {
         if (runningCount >= 500000000) {
